ultimate_quant_machine_v4_ultra/requirements.txt
pandas==2.1.4
numpy==1.24.3
python-binance==1.0.17
websocket-client==1.7.0
aiohttp==3.9.1
scipy==1.11.4
ta==0.10.2
pyyaml==6.0.1
requests==2.31.0
python-telegram-bot==20.0
plotly==5.18.0
ccxt==4.0.0
pytz==2023.3
